ajout de la carte pour le site vitrine
